refactor(api-server): extract reports helpers into helpers.rs
Behavior-preserving structural split of the reports churn hotspot
(backend/servers/api-server/src/routes/reports/mod.rs). Moves the four
internal helper functions (get_building_name, estimate_report_row_count,
generate_sync_csv_report, get_content_type_for_format) into a new
helpers.rs submodule. No routing, handler, or response-shape changes.

mod.rs: 2460 -> 2185 lines.
